requirements.txt

streamlit==1.26.0
pandas==2.1.1
numpy==1.26.2
matplotlib==3.8.0
seaborn==0.12.2
plotly==5.20.0
mysql-connector-python==8.1.1
pymysql==1.1.1

